Code refactor (no functional changes) (#3061)
* Code refactor (no functional changes)

* Update clients/daco-dialect-support/src/engine/copyfrom.ts


* Address comments

---------
