+++ clients/daco-dialect-support/src/engine/util.ts
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2026 Broadcom.
+ * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Broadcom - initial API and implementation
+ */
+import * as vscode from "vscode";
+import { IDocumentProcessingContext } from "@code4z/cobol-dialect-api";
+import { ParseError } from "./parsing";
+
+const FILLER_NAME = "FILLER";
+
+export function updateVariableName(name: string, suffix: string) {
+  console.log(`Updating variable name '${name}' with suffix '${suffix}'...`);
+  if (name.toUpperCase() === FILLER_NAME) {
+    return name;
+  }
+
+  const dashIndex = name.lastIndexOf("-");
+  if (dashIndex === name.length - 4) {
+    return name.substring(0, dashIndex + 2) + suffix;
+  }
+  return name + suffix;
+}
+
+export function extractSuffix(parentName: string | undefined): string {
+  if (!parentName) {
+    return "";
+  }
+  if (parentName.length > 2) {
+    return parentName.substring(parentName.length - 2);
+  }
+  return "";
+}
+
+export function addParsingErrors(
+  context: IDocumentProcessingContext,
+  errors: ParseError[],
+) {
+  errors.forEach((error) => {
+    context.addDiagnostic({
+      severity: vscode.DiagnosticSeverity.Error,
+      message: error.message,
+      range: error.range,
+    });
+  });
+}
